Output port names in kubectl port mappings
When --ports-file is used to output a mapping of ports from
kubectl port-forward, include the original port name or number
the client application requested. This allows the app to find
the port mappings that correspond to the ports it requested even
after service-port remapping and named port resolution is done.